Updated Alpha-Beta Pruning to return fastest win
Previously, the first best move found would be returned. Now, we're checking to make sure that any returned evaluations will have the best optimal evaluation AND depth.
If the AI knows that it'll win in the future, it may play some move that is guaranteed to win regardless of the player's next move. While this level of BM may be hilarious, it means that there's technically a better move on the board.